Major Refactoring
BindingTunnelKai's source selection was refactored to be cleaner.

BindingTunnel was deleted and BindingTunnelKai was changed to BindingTunnel

Binding was refactored to have a cleaner source selection

IsModeLocked and PrefferedMode replaced the old method based way of checking if two way binding (or some other mode was required). By making it property based it's now much more standard to query.

Binding now shows the binding mode in a disabled state if the mode is locked.

Serialization callbacks are now being used by the Binding Script to simplify Editor to Binding component translation.

There is an EditorHelper class now. It currently only has a single purpose. It can create a menu to select a list of components given an object property. It's used a few times to make better component wiring. The driver script (which originally had these methods is now wired to the EditorHelper as well.