Improvements and Minor Fixing
Creation of a new mediator script that doubles as a binding source. The new system takes more steps that the binding script I made earlier, but it is easier to use with things like drivers and seems to be a little more versatile. Like drivers there will be different mediators for different types of display values.

Fixes to some of the dirver Editor scripts that was causing offset properties to not apply changes.