Fixed the average value settings for existing drivers
The existing drivers (excluding boolean and string) should now be correctly summing or averaging the source values. Previously the value was ignored for almost all drivers.

Note that offset is applied AFTER any averaging or summing. This feature is not tested, but it is unlikely that any new bugs were introduced.