Creation of Vector2 to Float Driver
Created a driver that takes a Vector2 value and routes parts of it to a single float. It can sum up multiple values, scale each x or y value by a constant value etc. It usually is used to just take a single value (x or y) and set a float based on it.