Reintroduciton of a method in the mediator class to focus on the input field
The UI mediator has a method to focus on the first input field it is connected to. It previously wasn't working with text mesh pro input fields. After remaking the components in Unity it started to work with no changes on my part implying that something to do with layout was causing issues.