Update MediatorUI.cs
Fix that makes it so a mediator checks if initialization has occurred before it sets the value. This makes it so the value of a mediator can be changed before Start is called. This was done so a prefab with a mediator can be instantiated and the value set using the mediator (instead of having to think about what UI component is attached)