Drivers can now target scriptable objects
The target variable in drivers was changed from a component to a UnityEngine.Object which is more general. Don't think this will cause any problems since I was just using general reflection on the target anyway.

I also changed the binding source SO to use OnEnable() instead of OnAwake(). These functions are not super consistent so I don't have confidence they will work. One user suggested calling the SO from a Monobehaviors events in some sort of manager. Looks like it could be worth looking into.